Fixed creating signer token
